codes/requirements.txt
apptools==5.3.0
asttokens==2.4.1
cffi==1.17.1
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.11
dcor==0.6
ddeint==0.3.0
debugpy==1.8.7
decorator==5.1.1
envisage==7.0.3
executing==2.1.0
fdasrsf==2.6.1
findiff==0.10.0
fonttools==4.46.0
ijson==3.3.0
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.43.0
matplotlib==3.8.2
matplotlib-inline==0.1.7
mayavi==4.8.2
mpmath==1.3.0
multimethod==1.12
nest-asyncio==1.6.0
numba==0.60.0
numpy==1.26.2
packaging==23.2
pandas==2.1.4
parso==0.8.4
patsy==0.5.6
Pillow==10.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.22
pyface==8.0.0
Pygments==2.18.0
pyparsing==3.1.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==308
PyYAML==6.0.1
pyzmq==26.2.0
rdata==0.11.2
scikit-datasets==0.2.4
scikit-fda==0.9.1
scikit-learn==1.5.2
scipy==1.11.4
six==1.16.0
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
traits==6.4.3
traitsui==8.0.0
typing_extensions==4.12.2
tzdata==2023.3
vtk==9.3.1
wcwidth==0.2.13
xarray==2024.9.0
